Renamed master to leader
